asdg

+++ code/modules/tgui/tgui_input_list.dm
@@ -23,43 +23,14 @@
 			return
 	/// Client does NOT have tgui_input on: Returns regular input
 	// if(!user.client.prefs.read_preference(/datum/preference/toggle/tgui_input))
-	// 	return input(user, message, title, default) as null|anything in items
+	if(!usr.client.prefs.tgui_input_mode)
+		return input(user, message, title, default) as null|anything in items
 	var/datum/tgui_list_input/input = new(user, message, title, items, default, timeout)
 	input.ui_interact(user)
 	input.wait()
 	if (input)
 		. = input.choice
 		qdel(input)
-
-/**
- * Creates an asynchronous TGUI input list window with an associated callback.
- *
- * This proc should be used to create inputs that invoke a callback with the user's chosen option.
- * Arguments:
- * * user - The user to show the input box to.
- * * message - The content of the input box, shown in the body of the TGUI window.
- * * title - The title of the input box, shown on the top of the TGUI window.
- * * items - The options that can be chosen by the user, each string is assigned a button on the UI.
- * * default - If an option is already preselected on the UI. Current values, etc.
- * * callback - The callback to be invoked when a choice is made.
- * * timeout - The timeout of the input box, after which the menu will close and qdel itself. Set to zero for no timeout.
- */
-/proc/tgui_input_list_async(mob/user, message, title = "Select", list/items, default, datum/callback/callback, timeout = 60 SECONDS)
-	if (!user)
-		user = usr
-	if(!length(items))
-		return
-	if (!istype(user))
-		if (istype(user, /client))
-			var/client/client = user
-			user = client.mob
-		else
-			return
-	/// Client does NOT have tgui_input on: Returns regular input
-	// if(!user.client.prefs.read_preference(/datum/preference/toggle/tgui_input))
-	// 	return input(user, message, title) as null|anything in items
-	var/datum/tgui_list_input/async/input = new(user, message, title, items, default, callback, timeout)
-	input.ui_interact(user)
 
 /**
  * # tgui_list_input
@@ -95,7 +66,7 @@
 	src.default = default
 	var/list/repeat_items = list()
 
-	// Gets rid of illegal characters
+	// Gets rid of illegal characters.
 	var/static/regex/whitelistedWords = regex(@{"([^\u0020-\u8000]+)"})
 
 	for(var/i in items)
@@ -103,10 +74,8 @@
 			continue
 
 		var/string_key = whitelistedWords.Replace("[i]", "")
-
-		//avoids duplicated keys E.g: when areas have the same name
+		// Avoids duplicated keys E.g: when areas have the same name.
 		string_key = avoid_assoc_duplicate_keys(string_key, repeat_items)
-
 		src.items += string_key
 		src.items_map[string_key] = i
 
@@ -142,18 +111,20 @@
 	return GLOB.always_state
 
 /datum/tgui_list_input/ui_static_data(mob/user)
-	. = list()
-	.["init_value"] = default || items[1]
-	.["items"] = items
-	.["large_buttons"] = FALSE//user.client.prefs.read_preference(/datum/preference/toggle/tgui_input_large)
-	.["message"] = message
-	.["swapped_buttons"] = FALSE//user.client.prefs.read_preference(/datum/preference/toggle/tgui_input_swapped)
-	.["title"] = title
+	var/list/data = list()
+	data["init_value"]      = default || items[1]
+	data["items"]           = items
+	data["large_buttons"]   = user.client.prefs.tgui_large_buttons     //user.client.prefs.read_preference(/datum/preference/toggle/tgui_input_large)
+	data["message"]         = message
+	data["swapped_buttons"] = !user.client.prefs.tgui_swapped_buttons  //user.client.prefs.read_preference(/datum/preference/toggle/tgui_input_swapped)
+	data["title"]           = title
+	return data
 
 /datum/tgui_list_input/ui_data(mob/user)
-	. = list()
+	var/list/data = list()
 	if(timeout)
-		.["timeout"] = clamp((timeout - (world.time - start_time) - 1 SECONDS) / (timeout - 1 SECONDS), 0, 1)
+		data["timeout"] = clamp((timeout - (world.time - start_time) - 1 SECONDS) / (timeout - 1 SECONDS), 0, 1)
+	return data
 
 /datum/tgui_list_input/ui_act(action, list/params)
 	. = ..()
@@ -174,6 +145,37 @@
 
 /datum/tgui_list_input/proc/set_choice(choice)
 	src.choice = choice
+
+/**
+ * Creates an asynchronous TGUI input list window with an associated callback.
+ *
+ * This proc should be used to create inputs that invoke a callback with the user's chosen option.
+ * Arguments:
+ * * user - The user to show the input box to.
+ * * message - The content of the input box, shown in the body of the TGUI window.
+ * * title - The title of the input box, shown on the top of the TGUI window.
+ * * items - The options that can be chosen by the user, each string is assigned a button on the UI.
+ * * default - If an option is already preselected on the UI. Current values, etc.
+ * * callback - The callback to be invoked when a choice is made.
+ * * timeout - The timeout of the input box, after which the menu will close and qdel itself. Set to zero for no timeout.
+ */
+/proc/tgui_input_list_async(mob/user, message, title = "Select", list/items, default, datum/callback/callback, timeout = 60 SECONDS)
+	if (istext(user))
+		stack_trace("tgui_alert() received text for user instead of mob")
+		return
+	if(!user)
+		user = usr
+	if(!length(items))
+		return
+	if(!istype(user))
+		if (istype(user, /client))
+			var/client/client = user
+			user = client.mob
+		else
+			return
+
+	var/datum/tgui_list_input/async/input = new(user, message, title, items, default, callback, timeout)
+	input.ui_interact(user)
 
 /**
  * # async tgui_list_input
